+++ package.json
@@ -22,11 +22,12 @@
     }
   },
   "dependencies": {
+    "gh-pages": "^2.0.1",
     "react": "next",
     "react-dom": "next",
     "react-router-dom": "^4.3.1",
     "styled-components": "^4.1.3",
-    "gh-pages": "^2.0.1"
+    "views": "^1.0.2"
   },
   "devDependencies": {
     "@babel/core": "^7.2.2",
