+++ src/components/common/icons/Icons.jsx
@@ -1,0 +1,8 @@
+import { Circle } from "./Circle.jsx";
+import { Place } from "./Place.jsx";
+import { Style } from "./Style.jsx";
+import { Date } from "./Date.jsx";
+import { Partner } from "./Partner.jsx";
+import { Comment } from "./Comment.jsx";
+
+export { Circle, Place, Style, Date, Partner, Comment };
