+++ src/components/views/logbook/Results.jsx
@@ -1,18 +1,19 @@
+import React from "react";
 import styled from "styled-components";
-import { colors, spacing, fonts, fontSize, fontWeight } from "./styleVars";
+import { colors, spacing, fonts, fontSize, fontWeight } from "../../common/styleVars";
+import { buttonBase } from "../../common/Buttons.jsx";
 
-const ListStyle = styled.ul`
+const ResultsList = styled.ul`
   margin-top: ${spacing.xLarge};
   li {
     width: 100%;
   }
 `;
 
+// todo:
+// - split up this button, see if spans in search results can be aligned to these
 const ListButton = styled.button`
-  user-select: none;
-  cursor: pointer;
-  transition: all ease-in-out 0.3s;
-  font-family: ${fonts.main};
+  ${buttonBase};
   font-size: ${fontSize.small};
   display: flex;
   align-items: center;
@@ -25,9 +26,6 @@ const ListButton = styled.button`
   &:hover {
     background-color: ${colors.lightGrey};
     border-color: ${colors.midGrey};
-  }
-  &:focus {
-    outline: none;
   }
   span:nth-child(1) {
     margin-right: ${spacing.large};
@@ -62,4 +60,29 @@ const ListButton = styled.button`
   }
 `;
 
-export { ListStyle, ListButton };
+const Results = ({ logs, low, high, onClick }) => {
+  return (
+    <ResultsList>
+      {logs.slice(low, high).map(log => (
+        <li key={log.key}>
+          <ListButton onClick={() => onClick(log.key)}>
+            <span>
+              <strong>{log.climbName}</strong> - {log.grade}
+            </span>{" "}
+            <span>
+              {log.style} - {log.cragName}
+            </span>
+            <span>
+              {log.date}
+              <svg xmlns="http://www.w3.org/2000/svg" width="48" height="48" viewBox="0 0 48 48">
+                <path d="M17.17 32.92l9.17-9.17-9.17-9.17L20 11.75l12 12-12 12z" />
+              </svg>
+            </span>
+          </ListButton>
+        </li>
+      ))}
+    </ResultsList>
+  );
+};
+
+export default Results;
