+++ src/index.html
@@ -8,7 +8,7 @@
     <meta name="description" content="A logbook for my climbs" />
     <link href="https://fonts.googleapis.com/css?family=Raleway:400,500,600" rel="stylesheet" />
     <link rel="stylesheet" type="text/css" href="./reset.css" />
-    <link rel="shortcut icon" type="image/x-icon" href="icons/favicon.ico" />
+    <link rel="shortcut icon" type="image/x-icon" href="public/favicon.ico" />
   </head>
 
   <body>
