+++ src/index.js
@@ -1,5 +1,5 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./Components/App";
+import App from "./components/App.jsx";
 
 ReactDOM.render(<App />, document.getElementById("root"));
